Moved TriTiltrotor and GenericFixedWing setups to AeroBodyParams
